requirements.txt
# Google ADK and AI Models
google-adk>=1.0.0
google-genai>=0.7.0
litellm>=0.4.0

# API Clients
anthropic>=0.7.0
notion-client>=2.2.1
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
aiofiles>=23.2.1

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Async and Concurrency
asyncio
aiohttp>=3.9.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0